refactor: avatar types
