+++ projects/ion/src/lib/avatar/avatar.component.ts
@@ -3,14 +3,6 @@ import { AvatarType } from '../core/types/avatar';
 import { SizeType } from '../core/types/size';
 import { IconType } from '../icon/icon.component';
 
-export interface IonAvatarProps {
-  type: AvatarType;
-  size: SizeType;
-  value?: string;
-  image?: string;
-  onErrorImage?: string;
-}
-
 @Component({
   selector: 'ion-avatar',
   templateUrl: './avatar.component.html',
