+++ projects/ion/src/lib/core/types/avatar.ts
@@ -1,5 +1,15 @@
+import { SizeType } from './size';
+
 export enum AvatarType {
   initials = 'initials',
   icon = 'icon',
   photo = 'photo',
+}
+
+export interface IonAvatarProps {
+  type: AvatarType;
+  size: SizeType;
+  value?: string;
+  image?: string;
+  onErrorImage?: string;
 }
