+++ projects/ion/src/lib/core/types/index.ts
@@ -5,3 +5,4 @@ export * from './tooltip';
 export * from './card';
 export * from './tab-group';
 export * from './alert';
+export * from './avatar';
